update README - wget

+++ scripts/tomcat_installer.py
@@ -57,42 +57,42 @@ def cleanup():
 
 # mode: silent/prompt, c: continue/clean 
 def install(mode, c):
-  	version = get_version()
-  	if (mode == 'prompt'):
-  	   ans = raw_input ('Do you want to install Tomcat '  + version + '? [y/n] ') 
-  	   if (ans == 'n'):
+  version = get_version()
+  if (mode == 'prompt'):
+  	 ans = raw_input ('Do you want to install Tomcat '  + version + '? [y/n] ') 
+  	 if (ans == 'n'):
   		print 'Skip installing Tomcat'
-		exit()
-  	else:
-	   ans = 'y'
+		  exit()
+  	 else:
+	     ans = 'y'
 
-  	print 'Installing Tomcat version ' + version
-  	subprocess.call ('groupadd tomcat ', shell=True)
-  	subprocess.call ('useradd -s /bin/nologin -g tomcat -d /opt/tomcat tomcat ', shell=True)
-	major_version = version.split('.')[0]
-	tar_name = 'apache-tomcat-' + version 
-  	#subprocess.call ('wget http://archive.apache.org/dist/tomcat/tomcat-' + major_version + '/v' + version + '/bin/' + tar_name + '.tar.gz' , shell=True)
-  	subprocess.call ('tar -zxvf ' + tar_name + '.tar.gz'  + ' -C /opt ', shell=True)
-	subprocess.call (' rm -rf /opt/tomcat', shell=True)
-	subprocess.call (' ln -s /opt/' + tar_name + ' /opt/tomcat', shell=True)
-  	subprocess.call (' chgrp -R tomcat /opt/tomcat/conf ', shell=True)
-  	subprocess.call (' chmod g+rwx /opt/tomcat/conf ', shell=True)
-  	subprocess.call (' chmod g+r /opt/tomcat/conf/* ', shell=True)
-  	subprocess.call (' chown -R tomcat /opt/tomcat/logs/ /opt/tomcat/temp/ /opt/tomcat/webapps/ /opt/tomcat/work/ ', shell=True)
-  	subprocess.call (' chgrp -R tomcat /opt/tomcat/bin ', shell=True)
-  	subprocess.call (' chgrp -R tomcat /opt/tomcat/lib  ', shell=True)
-  	subprocess.call (' chmod g+rwx /opt/tomcat/bin ', shell=True)
-  	subprocess.call (' chmod g+r /opt/tomcat/bin/* ', shell=True)
+  print 'Installing Tomcat version ' + version
+  subprocess.call ('groupadd tomcat ', shell=True)
+  subprocess.call ('useradd -s /bin/nologin -g tomcat -d /opt/tomcat tomcat ', shell=True)
+  major_version = version.split('.')[0]
+  tar_name = 'apache-tomcat-' + version 
+  subprocess.call ('wget http://archive.apache.org/dist/tomcat/tomcat-' + major_version + '/v' + version + '/bin/' + tar_name + '.tar.gz' , shell=True)
+  subprocess.call ('tar -zxvf ' + tar_name + '.tar.gz'  + ' -C /opt ', shell=True)
+  subprocess.call (' rm -rf /opt/tomcat', shell=True)
+  subprocess.call (' ln -s /opt/' + tar_name + ' /opt/tomcat', shell=True)
+  subprocess.call (' chgrp -R tomcat /opt/tomcat/conf ', shell=True)
+  subprocess.call (' chmod g+rwx /opt/tomcat/conf ', shell=True)
+  subprocess.call (' chmod g+r /opt/tomcat/conf/* ', shell=True)
+  subprocess.call (' chown -R tomcat /opt/tomcat/logs/ /opt/tomcat/temp/ /opt/tomcat/webapps/ /opt/tomcat/work/ ', shell=True)
+  subprocess.call (' chgrp -R tomcat /opt/tomcat/bin ', shell=True)
+  subprocess.call (' chgrp -R tomcat /opt/tomcat/lib  ', shell=True)
+  subprocess.call (' chmod g+rwx /opt/tomcat/bin ', shell=True)
+  subprocess.call (' chmod g+r /opt/tomcat/bin/* ', shell=True)
   
 
  	# update server.xml and tomcat.service 
 	if (mode == 'silent'):
-	   server_config = get_server_config()
-           print 'Use server.xml: ' + server_config
-	   service_config = get_service_config()
-	   print 'Use tomcat.service: ' + service_config
-	   subprocess.call (' cp ' + server_config + ' ../apache-configs' , shell=True)
-	   subprocess.call (' cp ' + service_config + ' ../apache-configs', shell=True)
+    server_config = get_server_config()
+    print 'Use server.xml: ' + server_config
+	  service_config = get_service_config()
+    print 'Use tomcat.service: ' + service_config
+	  subprocess.call (' cp ' + server_config + ' ../apache-configs' , shell=True)
+	  subprocess.call (' cp ' + service_config + ' ../apache-configs', shell=True)
 	else:
   	   print ('Please update keystore path. Save changes when you are done..')
   	   sleep (5)
@@ -101,19 +101,17 @@ def install(mode, c):
   	   sleep (5)
   	   subprocess.call('vim +17 ../apache-configs/tomcat.service', shell=True) 	
   	# move the files to the right location
-  	subprocess.call (' cp ../apache-configs/server.xml /opt/tomcat/conf', shell=True)	
-  	subprocess.call (' cp ../apache-configs/tomcat.service /etc/systemd/system/tomcat.service', shell=True)	
+  subprocess.call (' cp ../apache-configs/server.xml /opt/tomcat/conf', shell=True)	
+  subprocess.call (' cp ../apache-configs/tomcat.service /etc/systemd/system/tomcat.service', shell=True)	
   
 	print 'starting tomcat service ....'
 	subprocess.call (' systemctl daemon-reload', shell=True)
-  	subprocess.call (' systemctl restart tomcat.service', shell=True)
-  	subprocess.call (' systemctl enable tomcat.service', shell=True)	
+  subprocess.call (' systemctl restart tomcat.service', shell=True)
+  subprocess.call (' systemctl enable tomcat.service', shell=True)	
 	sleep (10)
 
   	# finish
-  	print 'Tomcat Installation completed.'
-	#subprocess.call('rm apach*.tar.gz*', shell=True)
-
+	print 'Tomcat Installation completed.'
 
 if __name__ == '__main__':
 	mode = 'prompt'
